Add separate get requests code in another file
